test/libpmi: update PMI simple proto test
Move t/pmi/simple to libpmi/test/simple where it belongs.

Eliminate duplicate dgetline/dputline functions and use the
ones we added to libpmi.

Add checks for edge cases:
* barrier entry failrue
* barrier exit failure
* kvs_put value too long
* kvs_put key too long
* kvs_get key too long
* kvs_get unknown key

Use pmi_simple_client_*() functions rather than generic pmi_() ones.
