+++ t/Makefile.am
@@ -29,7 +29,6 @@ TESTS = \
 	loop/multrpc.t \
 	loop/reduce.t \
 	loop/log.t \
-	pmi/simple.t \
 	t0000-sharness.t \
 	t0001-basic.t \
 	t0002-request.t \
@@ -146,7 +145,6 @@ check_PROGRAMS = \
 	loop/reduce.t \
 	loop/log.t \
 	pmi/kvstest \
-	pmi/simple.t \
 	pmi/pminfo \
 	kz/kzutil \
 	kvs/torture \
@@ -224,10 +222,6 @@ loop_reduce_t_LDADD = $(test_ldadd) $(LIBDL)
 pmi_kvstest_SOURCES = pmi/kvstest.c
 pmi_kvstest_CPPFLAGS = $(test_cppflags)
 pmi_kvstest_LDADD = $(test_ldadd) $(LIBRT)
-
-pmi_simple_t_SOURCES = pmi/simple.c
-pmi_simple_t_CPPFLAGS = $(test_cppflags)
-pmi_simple_t_LDADD = $(test_ldadd) $(LIBRT)
 
 pmi_pminfo_SOURCES = pmi/pminfo.c
 pmi_pminfo_CPPFLAGS = $(test_cppflags)
